pyproject.toml
[project]
name = "tool-user"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "cloudinary>=1.44.1",
    "fastapi>=0.115.14",
    "gradio-client>=1.10.4",
    "langchain>=0.3.26",
    "langchain-openai>=0.3.27",
    "langgraph>=0.5.1",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.35.0",
    "wikipedia>=1.4.0",
]
